mv: Rename content script versions
