+++ src/entrypoints/content/noop.ts
@@ -1,0 +1,3 @@
+export function main () {
+    console.log('[UFC] noop.ts on', document.location.href)
+}
